Fixed warnings and win64 error. JIRA CXX-1722
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@46739 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/lds2/lds2_db.cpp
@@ -185,7 +185,7 @@ void CLDS2_Database::Create(void)
 
     // Create tables:
 
-    for (int i = 0; i < sizeof(kLDS2_CreateDB)/sizeof(kLDS2_CreateDB[0]); i++) {
+    for (size_t i = 0; i < sizeof(kLDS2_CreateDB)/sizeof(kLDS2_CreateDB[0]); i++) {
         conn.ExecuteSql(kLDS2_CreateDB[i]);
     }
 }
